+++ packages/web/src/app/shared/component-documentation/component-example/component-example-generator/component-example-generator.component.ts
@@ -19,6 +19,7 @@ export class ComponentExampleGeneratorComponent implements OnInit, AfterContentI
   @Input() inlineExample = false;
   @Input() accordionCustom = false;
   @Input() overflowY;
+  @Input() overflowX;
   @Input() alignedTop = false;
   @Input() height = '340';
   enableFilters = true;