+++ packages/components/components/elvis-tabs/CHANGELOG.md
@@ -1,7 +1,8 @@
-# Elvia Checkbox Changelog
+# Elvia Tabs Changelog
 
-## 1.0.0 (13.11.20)
-
+## 1.0.4 (27.10.21)
+- Add a general fade
+- Remove margin when scroll is used
 ### Breaking changes
 
 - Breaking change